Use clang-tidy (with one deliberate error for test)

+++ tests/test_prx_files/src/set_each_word_in_buffer.c
@@ -8,7 +8,7 @@
 void SetEachWordInBuffer(const uint32_t value) {
 	// We may end up setting many bytes beyond the end of the visible buffer but that's fine
     // as the emulator does not complain
-    for(size_t offset = 0; offset < BUFFER_WIDTH * SCREEN_HEIGHT; offset++) {
+    for(size_t offset = 0; offset < (size_t)BUFFER_WIDTH * (size_t)SCREEN_HEIGHT; offset++) {
         *((uint32_t*)VRAM_TOP + offset) = value; 
     }
 }
